Added bottom axis, changed x-scale max
Added a bottom axis which is visible at all times. In order to do this
I've split the chart into two svg's. One thats wrapped in a div and is
scrollable and one that is visible below that div.

In order to share the same width between the top and bottom axis I've
had to round the xaxis max value to 100 as percentages are divisible by
100.